refactoring: BaseResourceProviderMixin to base_resource_provider_mixin
